[FIX] : document 컴포넌트 이벤트 하위 컴포넌트로 이동
