+++ src/Component/Sidebar/Document/DocumentForm/documentForm.js
@@ -1,9 +1,11 @@
 import { createNewElement } from '../../../../Util/element.js';
+import Observer from '../../../../Store/userObserver.js';
+import { request } from '../../../../Service/document.js';
 
 export default class DocumentForm {
-    constructor({ $target, onSubmitPost }) {
+    constructor({ $target, onSubmitCallback }) {
         this.$target = $target;
-        this.onSubmitPost = onSubmitPost;
+        this.onSubmitCallback = onSubmitCallback;
         this.$documentForm = createNewElement('form', [{ property: 'className', value: 'title-and-form' }]);
 
         this.init();
@@ -27,12 +29,26 @@ export default class DocumentForm {
         this.$documentForm.appendChild($button);
     }
 
+    // document 추가 핸들러
     handleOnSubmit(e) {
         e.preventDefault();
         const { className } = e.target;
 
         if (className === 'title-and-form') {
-            this.onSubmitPost(null);
+            this.postDocument(null);
+        }
+    }
+
+    // document 데이터 추가하기 API
+    async postDocument(parentId, title = '문서 제목') {
+        const currentUser = Observer.getInstance().getState();
+        const res = await request('/documents', currentUser, {
+            method: 'POST',
+            body: JSON.stringify({ title, parent: parentId }),
+        });
+
+        if (res) {
+            this.onSubmitCallback();
         }
     }
 }
